requirements_cpu_only.txt
python-docx==1.1.2
streamlit==1.40.1
openai==1.54.5
requests==2.32.3
python-dotenv==1.0.1